+++ tiger-admin/src/main/java/de/gematik/test/tiger/admin/helper/exceptions/ExceptionHandler.java
@@ -1,6 +1,7 @@
 package de.gematik.test.tiger.admin.helper.exceptions;
 
 import de.gematik.test.tiger.common.config.TigerConfigurationException;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
@@ -8,8 +9,6 @@ import org.json.JSONObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
-
-import java.io.IOException;
 
 @ControllerAdvice
 @Slf4j