+++ tiger-admin/src/test/java/de/gematik/test/tiger/admin/bdd/drivers/Parallel1IT.java
@@ -1,0 +1,13 @@
+package de.gematik.test.tiger.admin.bdd.drivers;
+
+import io.cucumber.junit.CucumberOptions;
+import net.serenitybdd.cucumber.CucumberWithSerenity;
+import org.junit.runner.RunWith;
+
+@RunWith(CucumberWithSerenity.class)
+@CucumberOptions(
+    features = {"./src/test/resources/features"},
+    plugin = {"json:target/cucumber-parallel/1.json"},
+    glue = { "de.gematik.test.tiger.admin.bdd.steps" })
+public class Parallel1IT {
+}