+++ tiger-test-lib/src/test/java/drivers/Parallel1IT.java
@@ -1,0 +1,13 @@
+package drivers;
+
+import io.cucumber.junit.CucumberOptions;
+import net.serenitybdd.cucumber.CucumberWithSerenity;
+import org.junit.runner.RunWith;
+
+@RunWith(CucumberWithSerenity.class)
+@CucumberOptions(
+    features = {"./src/test/resources/features/tiger-test-lib/testcontext.feature"},
+    plugin = {"json:target/cucumber-parallel/1.json"},
+    glue = { "de.gematik.test.tiger.glue" })
+public class Parallel1IT {
+}