+++ CHANGELOG.md
@@ -1,17 +1,26 @@
 # Changelog Tiger Testplattform
 
+## Release 0.16.3
+
+## Bugfix
+* TGR-219 bdd driver generator plugin created invalid feature path on windows 
+
+## Release 0.16.2
+
+## Bugfix
+* TGR-212 Content-Length now correctly after Rbel-Modifications
+
 ## Release 0.16.0
 
 ## Features
-* TGR-136 Client-Adressen werden nun in Rbel-Nachrichten korrekt gesetzt
-* TGR-120: Die SSL-Suiten des Servers sind nun konfigurierbar
+* TGR-136 Client-addresses are now correctly set in Rbel-messages
+* TGR-120 The SSL-Suits of the servers are now configurable
 
 ## Release 0.15.0
 
 ### Features
-* TGR-136 Client-Adressen werden nun in Rbel-Nachrichten korrekt gesetzt
+* TGR-136 Client-addresses are now correctly set in Rbel-Messages
 * TGR-186 First version of an UI test run monitor, displaying all banner and text messages to guide manual testers.
-* TGR-136 Client-Adressen werden nun in Rbel-Nachrichten korrekt gesetzt
 
 ### Breaking Changes